test: model test code 추가
